Animated first three elements
Made the show/hide button show and hide the box using a click function
in js. Created a class in css and added to an existing class.
Made the move in/out button move the diamond in and out by using a
click function in js. Created a class in css and added to two existing
classes.
Made the collapse/expand button collapse and expand the box using a
click function in js. Created a class in css and added to an existing
class.